tests/wpt/web-platform-tests/navigation-api/scroll-restoration/after-transition-timing.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<div style="height: 1000px; width: 1000px;"></div>
<div id="frag"></div>
<script>
promise_test(async t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  await new Promise(resolve => window.onload = () => t.step_timeout(resolve, 0));
  assert_equals(window.scrollY, 0);
  await navigation.navigate("#frag").finished;
  assert_not_equals(window.scrollY, 0);

  let transitionWhile_resolve;
  navigation.onnavigate = t.step_func(e => {
    e.transitionWhile(new Promise(r => transitionWhile_resolve = r));
    assert_not_equals(window.scrollY, 0);
  }), { scrollRestoration: "after-transition" };

  window.onpopstate = t.step_func(() => assert_not_equals(window.scrollY, 0));
  window.onhashchange = t.step_func(() => assert_not_equals(window.scrollY, 0));

  // The scroll restore should take place before navigatesuccess fires.
  let navigatesuccess_called = false;
  navigation.onnavigatesuccess = t.step_func(() => {
    navigatesuccess_called = true;
    assert_equals(window.scrollY, 0);
  });

  let back_promises = navigation.back();

  await back_promises.committed;
  transitionWhile_resolve();
  assert_not_equals(window.scrollY, 0);

  await back_promises.finished;
  assert_equals(window.scrollY, 0);
  assert_true(navigatesuccess_called);
}, "scrollRestoration: after-transition should scroll when back completes, just before navigatesuccess");
</script>
</body>
